requirements.txt
celery == 5.6.2

fastapi==0.110.0
uvicorn==0.27.1


sqlalchemy==2.0.48
asyncpg==0.29.0
alembic==1.13.1



redis==5.0.1


aiohttp==3.9.3


pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1


greenlet==3.3.2



httpx==0.27.0
pytest-asyncio==1.3.0